+++ src/mira/core/passes.py
@@ -14,16 +14,13 @@ from mira.config import load_config
 from mira.dashboard.models_config import llm_config_for
 from mira.exceptions import ResponseParseError
 from mira.llm.prompts.review import build_security_review_prompt
-from mira.llm.provider import (
-    SUBMIT_CRITIQUE_TOOL,
-    SUBMIT_REVIEW_TOOL,
-    LLMProvider,
-)
+from mira.llm.provider import LLMProvider
 from mira.llm.response_parser import (
     convert_to_review_comments,
     loads_lenient,
     parse_llm_response,
 )
+from mira.llm.tool_schemas import SUBMIT_CRITIQUE_TOOL, SUBMIT_REVIEW_TOOL
 from mira.models import KeyIssue, ReviewComment, Severity
 
 logger = logging.getLogger(__name__)
