+++ src/mira/llm/provider.py
@@ -2,8 +2,8 @@
 
 Per-provider quirks (attribution headers, model-prefix policy, reasoning
 remapping) come from the profile registry in ``mira.llm.providers``, matched
-to the configured ``base_url``. OpenRouter is one profile among many; any
-OpenAI-compatible endpoint works, with or without a profile entry.
+to the configured ``base_url``. OpenRouter is the one profile with quirks; any
+other OpenAI-compatible endpoint works off the portable default, no entry needed.
 """
 
 from __future__ import annotations
@@ -18,260 +18,9 @@ from tenacity import retry, retry_if_exception_type, stop_after_attempt, wait_ex
 from mira.config import LLMConfig
 from mira.exceptions import LLMError
 from mira.llm import providers
+from mira.llm.tool_schemas import SUBMIT_REVIEW_TOOL, SUBMIT_WALKTHROUGH_TOOL
 
 logger = logging.getLogger(__name__)
-
-
-SUBMIT_REVIEW_TOOL = {
-    "type": "function",
-    "function": {
-        "name": "submit_review",
-        "description": "Submit your code review findings including comments, key issues, and a summary.",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "comments": {
-                    "type": "array",
-                    "description": "List of review comments on specific lines of code.",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "path": {"type": "string", "description": "Relative file path."},
-                            "line": {
-                                "type": "integer",
-                                "description": "Line number in the target file.",
-                            },
-                            "end_line": {
-                                "type": ["integer", "null"],
-                                "description": "End line for multi-line comments, or null.",
-                            },
-                            "severity": {
-                                "type": "string",
-                                "enum": ["blocker", "warning", "suggestion", "nitpick"],
-                            },
-                            "category": {
-                                "type": "string",
-                                "enum": [
-                                    "bug",
-                                    "security",
-                                    "performance",
-                                    "error-handling",
-                                    "race-condition",
-                                    "resource-leak",
-                                    "maintainability",
-                                    "clarity",
-                                    "configuration",
-                                    "other",
-                                ],
-                            },
-                            "title": {"type": "string", "description": "Short title (<80 chars)."},
-                            "body": {
-                                "type": "string",
-                                "description": "Detailed explanation of the issue. Use single backticks for inline code references. Do NOT use triple-backtick code blocks.",
-                            },
-                            "confidence": {"type": "number", "minimum": 0.0, "maximum": 1.0},
-                            "existing_code": {
-                                "type": "string",
-                                "description": "Verbatim copy of the code from the diff that this comment targets. Must be an exact substring.",
-                            },
-                            "suggestion": {
-                                "type": ["string", "null"],
-                                "description": "Optional replacement code to fix the issue. Raw code only — do NOT wrap in backticks or markdown fences.",
-                            },
-                            "agent_prompt": {
-                                "type": ["string", "null"],
-                                "description": "Concise imperative instruction for AI coding agents.",
-                            },
-                        },
-                        "required": [
-                            "path",
-                            "line",
-                            "severity",
-                            "category",
-                            "title",
-                            "body",
-                            "confidence",
-                            "existing_code",
-                        ],
-                    },
-                },
-                "key_issues": {
-                    "type": "array",
-                    "description": "1-3 most critical findings a human reviewer MUST examine.",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "issue": {"type": "string"},
-                            "path": {"type": "string"},
-                            "line": {"type": "integer"},
-                        },
-                        "required": ["issue", "path", "line"],
-                    },
-                },
-                "summary": {
-                    "type": "string",
-                    "description": "Brief overall summary of the review.",
-                },
-                "metadata": {
-                    "type": "object",
-                    "properties": {
-                        "reviewed_files": {"type": "integer"},
-                        "skipped_reason": {"type": ["string", "null"]},
-                    },
-                },
-            },
-            "required": ["comments", "summary"],
-        },
-    },
-}
-
-SUBMIT_CRITIQUE_TOOL = {
-    "type": "function",
-    "function": {
-        "name": "submit_critique",
-        "description": (
-            "For each draft review comment, grade how well the evidence in "
-            "the diff supports the claimed issue."
-        ),
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "verdicts": {
-                    "type": "array",
-                    "description": "One verdict per draft comment, in input order.",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "index": {
-                                "type": "integer",
-                                "description": "Zero-based index of the draft comment.",
-                            },
-                            "evidence": {
-                                "type": "string",
-                                "enum": ["proven", "plausible", "unsupported"],
-                                "description": (
-                                    "proven: the shown code demonstrates the issue and the "
-                                    "reasoning is correct. "
-                                    "plausible: the issue is consistent with the shown code "
-                                    "but depends on behaviour or code not shown — this is a "
-                                    "valid grade for real findings, not a failure. "
-                                    "unsupported: the shown code contradicts the claim, the "
-                                    "reasoning is wrong, or it's a style preference dressed "
-                                    "up as an issue."
-                                ),
-                            },
-                            "reason": {
-                                "type": "string",
-                                "description": "One short sentence explaining the verdict.",
-                            },
-                        },
-                        "required": ["index", "evidence", "reason"],
-                    },
-                },
-            },
-            "required": ["verdicts"],
-        },
-    },
-}
-
-
-SUBMIT_THREAD_REPLY_TOOL = {
-    "type": "function",
-    "function": {
-        "name": "submit_thread_reply",
-        "description": (
-            "Reply to a human's comment on one of your previous PR review "
-            "suggestions. Classify their intent and write a short reply."
-        ),
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "intent": {
-                    "type": "string",
-                    "enum": ["disagreement", "question", "agreement", "other"],
-                    "description": (
-                        "disagreement = human refutes the suggestion / says it doesn't apply. "
-                        "question = human is asking for clarification. "
-                        "agreement = human is acknowledging or thanking. "
-                        "other = anything else (off-topic, unclear)."
-                    ),
-                },
-                "reply": {
-                    "type": "string",
-                    "description": (
-                        "Your reply, 1-2 short sentences, plain text, no markdown. "
-                        'No emojis, no apologies, no "as an AI". For disagreement, '
-                        "concede gracefully. For questions, answer directly."
-                    ),
-                },
-            },
-            "required": ["intent", "reply"],
-        },
-    },
-}
-
-
-SUBMIT_WALKTHROUGH_TOOL = {
-    "type": "function",
-    "function": {
-        "name": "submit_walkthrough",
-        "description": "Submit a high-level walkthrough summary of the pull request.",
-        "parameters": {
-            "type": "object",
-            "properties": {
-                "summary": {"type": "string", "description": "Brief overall summary of the PR."},
-                "confidence_score": {
-                    "type": "object",
-                    "properties": {
-                        "score": {"type": "integer", "minimum": 1, "maximum": 5},
-                        "label": {"type": "string"},
-                        "reason": {"type": "string"},
-                    },
-                    "required": ["score", "label", "reason"],
-                },
-                "change_groups": {
-                    "type": "array",
-                    "items": {
-                        "type": "object",
-                        "properties": {
-                            "label": {"type": "string"},
-                            "files": {
-                                "type": "array",
-                                "items": {
-                                    "type": "object",
-                                    "properties": {
-                                        "path": {"type": "string"},
-                                        "change_type": {
-                                            "type": "string",
-                                            "enum": ["added", "modified", "deleted", "renamed"],
-                                        },
-                                        "description": {"type": "string"},
-                                    },
-                                    "required": ["path", "change_type", "description"],
-                                },
-                            },
-                        },
-                        "required": ["label", "files"],
-                    },
-                },
-                "effort": {
-                    "type": "object",
-                    "properties": {
-                        "level": {"type": "integer", "minimum": 1, "maximum": 5},
-                        "label": {"type": "string"},
-                        "minutes": {"type": "integer"},
-                    },
-                    "required": ["level", "label", "minutes"],
-                },
-                "sequence_diagram": {
-                    "type": ["string", "null"],
-                    "description": "Mermaid sequence diagram or null.",
-                },
-            },
-            "required": ["summary", "change_groups"],
-        },
-    },
-}
 
 
 def _get_api_key(config: LLMConfig, profile: dict | None = None) -> str:
